requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
loggly-python-handler==1.0.0
mqtthandler==0.1.3
paho-mqtt==1.3.1
pycurl==7.43.0
pygobject==3.20.0
python-apt==1.1.0b1
PyYAML==3.12
requests==2.18.4
requests-futures==0.9.7
unattended-upgrades==0.1
urllib3==1.22
